backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.25.0
pandas==2.1.1
openai==1.2.0
python-dotenv==1.0.0
alembic==1.12.0
pytest==7.4.2
